+++ CHANGELOG.md
@@ -35,9 +35,11 @@ Add `- unreleased` along the next future version to prevent CI/CD from triggerin
 - adds `--no-clean` flag to `orbit build` and `orbit test` to prevent automatic cleaning before build process
 
 ### Changes
+- source file visibility levels are ignored for relative ip dependencies
 - renames `orbit launch` to `orbit publish` in favor of clearer intentions behind the command
 
 ### Fixes
+- resolves bug with protected files not being detected during hdl source code analysis as a cached ip or relative ip dependency
 - resolves bug where string swapping would not occur on source url if using a custom protocol
 
 ### Docs